refine logging and analyze url

+++ main.py
@@ -1,24 +1,33 @@
 ﻿#!/usr/bin/env python
 import csv
 import gzip
+import re
 from argparse import ArgumentParser
 from collections import defaultdict
-from datetime import datetime, timezone
+from datetime import datetime, timezone, timedelta
 from pathlib import Path
 
 import boto3
 import dateutil.parser
 
 
-def print_progress(prefix, count, total=None):
-    print(
-        "{}... {:03.2f}{}".format(
-            prefix,
-            count * 100 / total if total else count,
-            '%' if total else ''
-        ), end='\r'
-    )
+class ProgressLogger:
+    prev_print_at = datetime.now()
 
+    def log(self, prefix, count, total=None):
+        if self.prev_print_at > datetime.now() - timedelta(seconds=1):
+            return
+        print(
+            "{}... {:03.2f}{}".format(
+                prefix,
+                count * 100 / total if total else count,
+                '%' if total else ''
+            ), end='\r'
+        )
+        self.prev_print_at = datetime.now()
+
+
+progress_logger = ProgressLogger()
 
 download_folder = Path("./download")
 
@@ -67,7 +76,7 @@ class LogDownloader:
                 str(download_folder / file_name)
             )
             count += 1
-            print_progress('Download', count, total)
+            progress_logger.log('Download', count, total)
         print("Download complete! {} existed files.".format(exist) + " " * 10)
 
 
@@ -76,7 +85,7 @@ merged_file = Path('./merged')
 
 def merge_logs():
     if merged_file.exists():
-        print("{} exists.".format(merged_file))
+        print("File, {}, exists. Skip merging logs.".format(merged_file))
         return
 
     log_paths = list(download_folder.glob('*.gz'))
@@ -87,7 +96,7 @@ def merge_logs():
             with gzip.open(p, 'rb') as in_f:
                 out_f.write(in_f.read())
                 count += 1
-                print_progress('Decompression', count, total)
+                progress_logger.log('Decompression', count, total)
     print("Decompression {} files complete!".format(total) + " " * 10)
 
 
@@ -95,7 +104,12 @@ parsed_file = Path('./parsed')
 
 
 def parse_logs():
+    if parsed_file.exists():
+        print("File, {}, exists. Skip parsing logs.".format(parsed_file))
+        return
+
     count = 0
+    total = sum(1 for _ in merged_file.open('r'))
     with merged_file.open('r') as in_f, parsed_file.open('w') as out_f:
         for line in in_f:
             split = line.split(' ')
@@ -103,21 +117,48 @@ def parse_logs():
                 datetime=split[1], method=split[12][1:], url=split[13])
             )
             count += 1
-            print_progress('Parseing', count)
-    print("Parse complete!" + " " * 10)
+            progress_logger.log('Parsing', count, total)
+    print("Parsing logs complete!" + " " * 10)
 
 
 class LogAnalyzer:
     def __init__(self):
         self.stat_file = Path('./stat.csv')
         self.count = 0
+        self.normalize_handler = {
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/\d*/achievements'):
+                "https://api.soocii.me:443/graph/v1.2/<id>/achievements",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/\d*/followees/count'):
+                "https://api.soocii.me:443/graph/v1.2/<id>/followees/count",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/\d*/followers/count'):
+                "https://api.soocii.me:443/graph/v1.2/<id>/followers/count",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/users/\d*/followees'):
+                "https://api.soocii.me:443/graph/v1.2/users/<id>/followees",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/users/\d*/followers'):
+                "https://api.soocii.me:443/graph/v1.2/users/<id>/followers",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/\d*/friendship'):
+                "https://api.soocii.me:443/graph/v1.2/<id>/friendship",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/\d*/posts'):
+                "https://api.soocii.me:443/graph/v1.2/<id>/posts",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/me/feed/\w*-\w*'):
+                "https://api.soocii.me:443/graph/v1.2/me/feed/<status_id>",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/\w*-\w*/comments'):
+                "https://api.soocii.me:443/graph/v1.2/<status_id>/comments",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/posts/\w*-\w*/likes'):
+                "https://api.soocii.me:443/graph/v1.2/posts/<status_id>/likes",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/\d*/pinned-posts'):
+                "https://api.soocii.me:443/graph/v1.2/<id>/pinned-posts",
+            re.compile(r'https://api\.soocii\.me:443/graph/v1\.2/me/posts/\w*-\w'):
+                "https://api.soocii.me:443/graph/v1.2/me/posts/<status_id>"
+        }
 
     def stat_api_calls(self, start, end):
         stats = defaultdict(lambda: 0)
+        total = sum(1 for _ in parsed_file.open('r'))
         with parsed_file.open('r') as in_f:
             for line in in_f:
                 self.count += 1
-                print_progress("Analyzing line counts", self.count)
+                progress_logger.log("Analyzing", self.count, total)
                 line = line.replace('\n', '')
                 log_at, method, url = self._parse_line(line)
 
@@ -136,6 +177,8 @@ class LogAnalyzer:
                 split = key.split(' ')
                 service, method, url = split[0], split[1], split[2]
                 writer.writerow([service, method, url, count])
+
+        print("Analyzing logs complete!" + " " * 12)
 
     def _parse_line(self, line):
         log_datetime, method, url = line.split(' ')
@@ -161,7 +204,11 @@ class LogAnalyzer:
         return service
 
     def _normalize_url(self, url):
-        return url.split('?')[0]
+        url = url.split('?')[0]
+        for ptn, endpoint in self.normalize_handler.items():
+            if ptn.match(url):
+                url = endpoint
+        return url
 
 
 if __name__ == '__main__':
@@ -170,6 +217,7 @@ if __name__ == '__main__':
         dt = dt.astimezone(timezone.utc)
         dt = dt.replace(tzinfo=None)
         return dt
+
 
     arg_parser = ArgumentParser(description="Analyze ALB logs by datetime duration")
     arg_parser.add_argument('start', type=convert_datetime_str, help="Start datetime")
